requirements.txt
streamlit==1.28.0
pandas==2.1.3
plotly==5.17.0
requests==2.31.0
websocket-client==1.6.3
numpy==1.26.2
